==> ScRNA-seq Processing/Step1.Read_10X_data.ipynb <==
library(Seurat) #seurat V5
#推荐使用seurat V4
library(dplyr)

#有云服务器的，可开启并运算，这里我用4个线程：
library(future)
library(qs)
# check the current active plan
plan()
# change the current plan to access parallelization
plan("multisession", workers =40)
plan()

#设置可用的内存
# options(future.globals.maxSize = 4 * 1024^3)
plan("sequential")

## 1.1 lappy循环依次读入数据

## 1.1 lappy循环依次读入数据
# Create each individual Seurat object for every sample
fileID = list.files("./1.10X_files/matrix/")
path = "./1.10X_files/matrix/"
fileID

# 运行lapply函数读入数据
seurat.list = lapply(fileID, function(file){
  seurat_data <- Read10X(data.dir = paste0(path, file))
  seurat_obj <- CreateSeuratObject(counts = seurat_data,
                                   min.cells = 0,
                                   min.features = 0,
                                   project = file)
  return(seurat_obj)
})

names(seurat.list) = fileID
seurat.list

## 1.2 merge多个seurat对象

## 1.2 merge多个seurat对象
# Create a merged Seurat object
merged_seurat <- merge(x = seurat.list[[1]],
                       y = seurat.list[-1],
                       add.cell.id = fileID)
merged_seurat 

## 1.3 表型数据载入

## 1.3 表型数据载入
merged_seurat$sampleID = merged_seurat$orig.ident

# 根据样本信息重命名编组
merged_seurat$group <- recode(merged_seurat$sampleID,
                              "GSM8504078" = "CRL",
                              "GSM8504079" = "CRL",
                              "GSM8504080" = "CRL",
                              "GSM8863170" = "CRL",
                              "GSM8504081" = "ALI",
                              "GSM8504082" = "ALI",
                              "GSM8504083" = "ALI",
                              "GSM8863172" = "ALI"
                              )

# 'GSM8504078''GSM8504079''GSM8504080''GSM8504081''GSM8504082''GSM8504083''GSM8863170''GSM8863172'

head(merged_seurat@meta.data)

## 1.4 保存一下

#### （1）数据保存

#qs速度快
#install.packages('qs')
library(qs)
system.time({
    qsave(merged_seurat,file = "./Outdata/Step1.RawCount_merged_seurat.qs") 
})

#### （2）数据读取

# qread速度很快
library(qs)
system.time({
    input.data = qread(file = "./Outdata/Step1.RawCount_merged_seurat.qs")
           })

### End
