==> ScRNA-seq Processing/Step3.Cell_frequency.ipynb <==
library(Seurat)
library(dplyr)
library(patchwork)
library(readr)
library(ggplot2)
library(qs)
library(ggsci)
library(ggpubr)

# 加载绘图函数
source("./source_R/custom_seurat_functions.R")
if(T){
  text.size = 12
  text.angle = 45
  text.hjust = 1
  legend.position = "right"
  mytheme <- theme(plot.title = element_text(size = text.size+2,color="black",hjust = 0.5),
                   axis.ticks = element_line(color = "black"),
                   axis.title = element_text(size = text.size,color ="black"), 
                   axis.text = element_text(size=text.size,color = "black"),
                   axis.text.x = element_text(angle = text.angle, hjust = text.hjust ), #,vjust = 0.5
                   panel.grid=element_blank(), # 去网格线
                   legend.position = legend.position,
                   legend.text = element_text(size= text.size),
                   legend.title= element_text(size= text.size)
  )
}

### Step1.读入数据

### 1.读入数据
seurat.data = qread(file = "./Outdata/Step3.Cluster_annotion.qs")
seurat.data

table(seurat.data$sampleID)

table(seurat.data$group)

#UMAP降维可视化
options(repr.plot.width = 6, repr.plot.height = 5)
p1 = DimPlot(seurat.data, 
        reduction = "umap",
        group.by = "celltype",
        label = T);p1

### Step2.统计细胞比例

table(seurat.data$sampleID, seurat.data$group)

table(seurat.data$celltype)

library(paletteer)

#### 2.1 百分比柱状图
## 配色 
color.use=c('B cells'='#1f77b4', 'Endothelial cells'='#ff7f0e','Epithelial cells'= '#279e68', 'Fibroblast'='#d62728', 
            # 'DCs'='#aa40fc', 'Erythroid cells'='#8c564b', 'Mast cells'='#e377c2', 
            'Platelets'= '#b5bd61', 'T cells'='#17becf','Macrophage'='#e377c2', 
            'Neutrophils'= '#98df8a','NK cells'='#ff9896', 'Smooth muscle cells'= '#c5b0d5','Other Myeloid cells'='#aec7e8')
## 按分组
options(repr.plot.width = 5, repr.plot.height = 5)
p1 = plot.clusters.group(data = seurat.data,
                         Group = "group",
                         legend.position = "none",
                         celltype.id = "celltype",
                         widths = c(2,1),
                         log = F,
                         text.size = 12,
                         order = F,
                         legend.title = "Cell type",
                         color = 1,
                         xlab = "",
                         cell.counts = F)&mytheme&
    scale_fill_manual(values = color.use);p1

## 按分组 + 细胞数量统计
options(repr.plot.width = 9, repr.plot.height = 4)
p2 = plot.clusters.group(data = seurat.data,
                         Group = "group",
                         legend.position = "top",
                         celltype.id = "celltype",
                         widths = c(2,1),
                         log = F,
                         text.size = 12,
                         order = T,
                         legend.title = "Cell type",
                         color = 1,
                         xlab = "",
                         cell.counts = T)& 
    scale_fill_manual(values = color.use);p2
pdf(file="frequency_CRL_ALI.pdf", width=9, height=4)
print(p2)
dev.off()
